create definition.ts files and simplify the .emb.ts
