+++ src/components/charts/bars/BarChartStackedPro/BarChartStackedPro.emb.ts
@@ -1,96 +1,8 @@
-import { Granularity, Value, loadData } from '@embeddable.com/core';
-import {
-  defineComponent,
-  definePreview,
-  EmbeddedComponentMeta,
-  Inputs,
-} from '@embeddable.com/react';
-import BarChartStackedPro from './index';
-import { inputs } from '../../../component.inputs.constants';
-import { previewData } from '../../../preview.data.constants';
-import { getDimensionWithGranularity } from '../../utils/granularity.utils';
+import { defineComponent } from '@embeddable.com/react';
+import { barChartStackedPro } from './definition';
 
-export const meta = {
-  name: 'BarChartStackedPro',
-  label: 'Bar Chart - Stacked',
-  category: 'Bar Charts',
-  inputs: [
-    inputs.dataset,
-    inputs.measure,
-    { ...inputs.dimensionWithGranularitySelectField, name: 'xAxis', label: 'X-axis' },
-    inputs.groupBy,
-    inputs.title,
-    inputs.description,
-    inputs.tooltip,
-    inputs.maxResults,
-    inputs.showLegend,
-    inputs.showTooltips,
-    { ...inputs.showValueLabels, defaultValue: false },
-    inputs.showLogarithmicScale,
-    inputs.xAxisLabel,
-    inputs.yAxisLabel,
-    inputs.reverseXAxis,
-    inputs.yAxisRangeMin,
-    inputs.yAxisRangeMax,
-    inputs.showTotalLabels,
-  ],
-  events: [
-    {
-      name: 'onBarClicked',
-      label: 'A bar is clicked',
-      properties: [
-        {
-          name: 'axisDimensionValue',
-          label: 'Clicked axis dimension value',
-          type: 'string',
-        },
-        {
-          name: 'groupingDimensionValue',
-          label: 'Clicked grouping dimension value',
-          type: 'string',
-        },
-      ],
-    },
-  ],
-} as const satisfies EmbeddedComponentMeta;
+export const preview = barChartStackedPro.preview;
 
-export const preview = definePreview(BarChartStackedPro, {
-  xAxis: previewData.dimension,
-  groupBy: previewData.dimensionGroup,
-  measure: previewData.measure,
-  results: previewData.results1Measure2Dimensions,
-  hideMenu: true,
-  setGranularity: () => {},
-});
+export const meta = barChartStackedPro.meta;
 
-type BarChartStackedProState = {
-  granularity?: Granularity;
-};
-
-export default defineComponent(BarChartStackedPro, meta, {
-  props: (
-    inputs: Inputs<typeof meta>,
-    [state, setState]: [BarChartStackedProState, (state: BarChartStackedProState) => void],
-  ) => {
-    const xAxisWithGranularity = getDimensionWithGranularity(inputs.xAxis, state?.granularity);
-
-    return {
-      ...inputs,
-      xAxis: xAxisWithGranularity,
-      setGranularity: (granularity) => setState({ granularity }),
-      results: loadData({
-        limit: inputs.maxResults,
-        from: inputs.dataset,
-        select: [xAxisWithGranularity, inputs.groupBy, inputs.measure],
-      }),
-    };
-  },
-  events: {
-    onBarClicked: (value) => {
-      return {
-        axisDimensionValue: value.axisDimensionValue ?? Value.noFilter(),
-        groupingDimensionValue: value.groupingDimensionValue ?? Value.noFilter(),
-      };
-    },
-  },
-});
+export default defineComponent(barChartStackedPro.Component, meta, barChartStackedPro.config);
