+++ src/components/charts/kpis/KpiChartNumberPro/KpiChartNumberPro.emb.ts
@@ -1,46 +1,8 @@
-import {
-  defineComponent,
-  definePreview,
-  EmbeddedComponentMeta,
-  Inputs,
-} from '@embeddable.com/react';
-import KpiChartNumberPro from './index';
-import { loadData } from '@embeddable.com/core';
-import { inputs } from '../../../component.inputs.constants';
-import { previewData } from '../../../preview.data.constants';
+import { defineComponent } from '@embeddable.com/react';
+import { kpiChartNumberPro } from './definition';
 
-export const meta = {
-  name: 'KpiChartNumberPro',
-  label: 'Kpi Chart - Number',
-  category: 'Kpi Charts',
-  inputs: [
-    inputs.dataset,
-    inputs.measure,
-    inputs.title,
-    inputs.description,
-    inputs.displayNullAs,
-    inputs.tooltip,
-    inputs.fontSize,
-  ],
-} as const satisfies EmbeddedComponentMeta;
+export const preview = kpiChartNumberPro.preview;
 
-export const preview = definePreview(KpiChartNumberPro, {
-  measure: {
-    ...previewData.measure,
-  },
-  results: previewData.results1Measure,
-  fontSize: 100,
-  hideMenu: true,
-});
+export const meta = kpiChartNumberPro.meta;
 
-export default defineComponent(KpiChartNumberPro, meta, {
-  props: (inputs: Inputs<typeof meta>) => {
-    return {
-      ...inputs,
-      results: loadData({
-        from: inputs.dataset,
-        select: [inputs.measure],
-      }),
-    };
-  },
-});
+export default defineComponent(kpiChartNumberPro.Component, meta, kpiChartNumberPro.config);
