+++ src/components/charts/lines/LineChartDefaultPro/definition.ts
@@ -1,0 +1,132 @@
+import {
+  DataResponse,
+  Dimension,
+  Granularity,
+  LoadDataRequest,
+  Value,
+  loadData,
+} from '@embeddable.com/core';
+import { definePreview, EmbeddedComponentMeta, Inputs } from '@embeddable.com/react';
+import Component from './index';
+import { LineChartProOptionsClickArg } from '../lines.utils';
+import { inputs } from '../../../component.inputs.constants';
+import { subInputs } from '../../../component.subinputs.constants';
+import { previewData } from '../../../preview.data.constants';
+import { getDimensionWithGranularity } from '../../utils/granularity.utils';
+
+const meta = {
+  name: 'LineChartDefaultPro',
+  label: 'Line Chart - Default',
+  category: 'Line Charts',
+  inputs: [
+    inputs.dataset,
+    {
+      ...inputs.measures,
+      inputs: [
+        ...inputs.measures.inputs,
+        { ...subInputs.boolean, name: 'fillUnderLine', label: 'Fill under line' },
+        {
+          ...subInputs.color,
+          name: 'lineColor',
+          label: 'Line color',
+        },
+        {
+          ...subInputs.boolean,
+          name: 'connectGaps',
+          label: 'Connect gaps',
+          defaultValue: true,
+        },
+        {
+          ...subInputs.boolean,
+          name: 'dashedLine',
+          label: 'Dashed line',
+          defaultValue: false,
+        },
+      ],
+    },
+    { ...inputs.dimensionWithGranularitySelectField, label: 'X-axis', name: 'xAxis' },
+    inputs.title,
+    inputs.description,
+    inputs.tooltip,
+    inputs.showLegend,
+    inputs.showTooltips,
+    inputs.showValueLabels,
+    inputs.showLogarithmicScale,
+    inputs.xAxisLabel,
+    inputs.yAxisLabel,
+    inputs.reverseXAxis,
+    inputs.yAxisRangeMin,
+    inputs.yAxisRangeMax,
+    inputs.maxResults,
+  ],
+  events: [
+    {
+      name: 'onLineClicked',
+      label: 'A line is clicked',
+      properties: [
+        {
+          name: 'axisDimensionValue',
+          label: 'Clicked axis dimension value',
+          type: 'string',
+        },
+      ],
+    },
+  ],
+} as const satisfies EmbeddedComponentMeta;
+
+export type LineChartDefaultProState = {
+  granularity?: Granularity;
+};
+
+const previewConfig = {
+  xAxis: previewData.dimension,
+  measures: [previewData.measure],
+  results: previewData.results1Measure1Dimension,
+  hideMenu: true,
+};
+
+const preview = definePreview(Component, previewConfig);
+
+const loadDataResultsArgs = (inputs: Inputs<typeof meta>, xAxis?: Dimension): LoadDataRequest => ({
+  limit: inputs.maxResults,
+  from: inputs.dataset,
+  select: [...inputs.measures, xAxis ?? inputs.xAxis],
+});
+
+const loadDataResults = (inputs: Inputs<typeof meta>, xAxis: Dimension): DataResponse =>
+  loadData(loadDataResultsArgs(inputs, xAxis));
+
+const events = {
+  onLineClicked: (value: LineChartProOptionsClickArg) => ({
+    axisDimensionValue: value.dimensionValue ?? Value.noFilter(),
+  }),
+};
+
+const props = (
+  inputs: Inputs<typeof meta>,
+  [state, setState]: [LineChartDefaultProState, (state: LineChartDefaultProState) => void],
+) => {
+  const xAxisWithGranularity = getDimensionWithGranularity(inputs.xAxis, state?.granularity);
+
+  return {
+    ...inputs,
+    xAxis: xAxisWithGranularity,
+    setGranularity: (granularity: Granularity) => setState({ granularity }),
+    results: loadDataResults(inputs, xAxisWithGranularity),
+  };
+};
+
+export const lineChartDefaultPro = {
+  Component,
+  meta,
+  preview,
+  previewConfig,
+  config: {
+    props,
+    events,
+  },
+  results: {
+    loadDataArgs: loadDataResultsArgs,
+    loadData: loadDataResults,
+  },
+} as const;
