+++ src/components/charts/lines/LineChartGroupedPro/LineChartGroupedPro.emb.ts
@@ -1,114 +1,8 @@
-import {
-  defineComponent,
-  definePreview,
-  EmbeddedComponentMeta,
-  Inputs,
-} from '@embeddable.com/react';
-import LineChartGroupedPro from './index';
-import { Granularity, loadData, Value } from '@embeddable.com/core';
-import { LineChartProOptionsClickArg } from '../lines.utils';
-import { inputs } from '../../../component.inputs.constants';
-import { previewData } from '../../../preview.data.constants';
-import { getDimensionWithGranularity } from '../../utils/granularity.utils';
+import { defineComponent } from '@embeddable.com/react';
+import { lineChartGroupedPro } from './definition';
 
-export const meta = {
-  name: 'LineChartGroupedPro',
-  label: 'Line Chart - Grouped',
-  category: 'Line Charts',
-  inputs: [
-    inputs.dataset,
-    {
-      ...inputs.measure,
-      inputs: [
-        ...inputs.measure.inputs,
-        {
-          ...inputs.boolean,
-          name: 'fillUnderLine',
-          label: 'Fill under line',
-          category: 'Component Settings',
-        },
-        {
-          ...inputs.boolean,
-          name: 'connectGaps',
-          label: 'Connect gaps',
-          defaultValue: true,
-          category: 'Component Settings',
-        },
-      ],
-    },
-    { ...inputs.dimensionWithGranularitySelectField, name: 'xAxis', label: 'X-axis' },
-    inputs.groupBy,
-    inputs.title,
-    inputs.description,
-    inputs.tooltip,
-    inputs.maxResults,
-    inputs.showLegend,
-    inputs.showTooltips,
-    inputs.showValueLabels,
-    inputs.showLogarithmicScale,
-    inputs.xAxisLabel,
-    inputs.yAxisLabel,
-    inputs.reverseXAxis,
-    inputs.yAxisRangeMin,
-    inputs.yAxisRangeMax,
-  ],
-  events: [
-    {
-      name: 'onLineClicked',
-      label: 'A line is clicked',
-      properties: [
-        {
-          name: 'axisDimensionValue',
-          label: 'Clicked axis dimension value',
-          type: 'string',
-        },
-        {
-          name: 'groupingDimensionValue',
-          label: 'Clicked grouping dimension value',
-          type: 'string',
-        },
-      ],
-    },
-  ],
-} as const satisfies EmbeddedComponentMeta;
+export const preview = lineChartGroupedPro.preview;
 
-export const preview = definePreview(LineChartGroupedPro, {
-  xAxis: previewData.dimension,
-  groupBy: previewData.dimensionGroup,
-  measure: previewData.measure,
-  results: previewData.results1Measure2Dimensions,
-  hideMenu: true,
-  setGranularity: () => {},
-});
+export const meta = lineChartGroupedPro.meta;
 
-type LineChartGroupedProState = {
-  granularity?: Granularity;
-};
-
-export default defineComponent(LineChartGroupedPro, meta, {
-  props: (
-    inputs: Inputs<typeof meta>,
-    [state, setState]: [LineChartGroupedProState, (state: LineChartGroupedProState) => void],
-  ) => {
-    const xAxisWithGranularity = getDimensionWithGranularity(inputs.xAxis, state?.granularity);
-
-    return {
-      ...inputs,
-      xAxis: xAxisWithGranularity,
-      setGranularity: (granularity) => setState({ granularity }),
-      results: loadData({
-        limit: inputs.maxResults,
-        from: inputs.dataset,
-        select: [xAxisWithGranularity, inputs.groupBy, inputs.measure],
-      }),
-    };
-  },
-  events: {
-    onLineClicked: (value: LineChartProOptionsClickArg) => {
-      return {
-        axisDimensionValue: value.dimensionValue || Value.noFilter(),
-        groupingDimensionValue: value.groupingDimensionValue || Value.noFilter(),
-      };
-    },
-  },
-});
+export default defineComponent(lineChartGroupedPro.Component, meta, lineChartGroupedPro.config);
