+++ src/components/charts/pies/DonutChartPro/DonutChartPro.emb.ts
@@ -1,68 +1,8 @@
-import { Value, loadData } from '@embeddable.com/core';
-import {
-  defineComponent,
-  definePreview,
-  EmbeddedComponentMeta,
-  Inputs,
-} from '@embeddable.com/react';
-import DonutChartPro from './index';
-import { inputs } from '../../../component.inputs.constants';
-import { previewData } from '../../../preview.data.constants';
+import { defineComponent } from '@embeddable.com/react';
+import { donutChartPro } from './definition';
 
-export const meta = {
-  name: 'DonutChartPro',
-  label: 'Donut Chart',
-  category: 'Pie Charts',
-  inputs: [
-    inputs.dataset,
-    inputs.measure,
-    inputs.dimension,
-    inputs.title,
-    inputs.description,
-    inputs.tooltip,
-    inputs.showLegend,
-    inputs.maxLegendItems,
-    inputs.showTooltips,
-    inputs.showValueLabels,
-  ],
-  events: [
-    {
-      name: 'onSegmentClick',
-      label: 'A segment is clicked',
-      properties: [
-        {
-          name: 'dimensionValue',
-          label: 'Clicked dimension',
-          type: 'string',
-        },
-      ],
-    },
-  ],
-} as const satisfies EmbeddedComponentMeta;
+export const preview = donutChartPro.preview;
 
-export const preview = definePreview(DonutChartPro, {
-  dimension: previewData.dimension,
-  measure: previewData.measure,
-  results: previewData.results1Measure1Dimension,
-  showValueLabels: false,
-  hideMenu: true,
-});
+export const meta = donutChartPro.meta;
 
-export default defineComponent(DonutChartPro, meta, {
-  props: (inputs: Inputs<typeof meta>) => {
-    return {
-      ...inputs,
-      results: loadData({
-        from: inputs.dataset,
-        select: [inputs.measure, inputs.dimension],
-      }),
-    };
-  },
-  events: {
-    onSegmentClick: (value) => {
-      return {
-        dimensionValue: value.dimensionValue || Value.noFilter(),
-      };
-    },
-  },
-});
+export default defineComponent(donutChartPro.Component, meta, donutChartPro.config);
