+++ src/components/charts/pies/PieChartPro/PieChartPro.emb.ts
@@ -1,68 +1,8 @@
-import { Value, loadData } from '@embeddable.com/core';
-import {
-  defineComponent,
-  definePreview,
-  EmbeddedComponentMeta,
-  Inputs,
-} from '@embeddable.com/react';
-import PieChartPro from './index';
-import { inputs } from '../../../component.inputs.constants';
-import { previewData } from '../../../preview.data.constants';
+import { defineComponent } from '@embeddable.com/react';
+import { pieChartPro } from './definition';
 
-export const meta = {
-  name: 'PieChartPro',
-  label: 'Pie Chart',
-  category: 'Pie Charts',
-  inputs: [
-    inputs.dataset,
-    inputs.measure,
-    inputs.dimension,
-    inputs.title,
-    inputs.description,
-    inputs.tooltip,
-    inputs.showLegend,
-    inputs.maxLegendItems,
-    inputs.showTooltips,
-    inputs.showValueLabels,
-  ],
-  events: [
-    {
-      name: 'onSegmentClick',
-      label: 'A segment is clicked',
-      properties: [
-        {
-          name: 'dimensionValue',
-          label: 'Clicked dimension',
-          type: 'string',
-        },
-      ],
-    },
-  ],
-} as const satisfies EmbeddedComponentMeta;
+export const preview = pieChartPro.preview;
 
-export const preview = definePreview(PieChartPro, {
-  dimension: previewData.dimension,
-  measure: previewData.measure,
-  results: previewData.results1Measure1Dimension,
-  showValueLabels: false,
-  hideMenu: true,
-});
+export const meta = pieChartPro.meta;
 
-export default defineComponent(PieChartPro, meta, {
-  props: (inputs: Inputs<typeof meta>) => {
-    return {
-      ...inputs,
-      results: loadData({
-        from: inputs.dataset,
-        select: [inputs.measure, inputs.dimension],
-      }),
-    };
-  },
-  events: {
-    onSegmentClick: (value) => {
-      return {
-        dimensionValue: value.dimensionValue || Value.noFilter(),
-      };
-    },
-  },
-});
+export default defineComponent(pieChartPro.Component, meta, pieChartPro.config);
