+++ src/components/charts/pies/PieChartPro/definition.ts
@@ -1,0 +1,80 @@
+import { DataResponse, LoadDataRequest, Value, loadData } from '@embeddable.com/core';
+import { definePreview, EmbeddedComponentMeta, Inputs } from '@embeddable.com/react';
+import Component from './index';
+import { inputs } from '../../../component.inputs.constants';
+import { previewData } from '../../../preview.data.constants';
+
+const meta = {
+  name: 'PieChartPro',
+  label: 'Pie Chart',
+  category: 'Pie Charts',
+  inputs: [
+    inputs.dataset,
+    inputs.measure,
+    inputs.dimension,
+    inputs.title,
+    inputs.description,
+    inputs.tooltip,
+    inputs.showLegend,
+    inputs.maxLegendItems,
+    inputs.showTooltips,
+    inputs.showValueLabels,
+  ],
+  events: [
+    {
+      name: 'onSegmentClick',
+      label: 'A segment is clicked',
+      properties: [
+        {
+          name: 'dimensionValue',
+          label: 'Clicked dimension',
+          type: 'string',
+        },
+      ],
+    },
+  ],
+} as const satisfies EmbeddedComponentMeta;
+
+const previewConfig = {
+  dimension: previewData.dimension,
+  measure: previewData.measure,
+  results: previewData.results1Measure1Dimension,
+  showValueLabels: false,
+  hideMenu: true,
+};
+
+const preview = definePreview(Component, previewConfig);
+
+const loadDataResultsArgs = (inputs: Inputs<typeof meta>): LoadDataRequest => ({
+  from: inputs.dataset,
+  select: [inputs.measure, inputs.dimension],
+});
+
+const loadDataResults = (inputs: Inputs<typeof meta>): DataResponse =>
+  loadData(loadDataResultsArgs(inputs));
+
+const events = {
+  onSegmentClick: (value: { dimensionValue?: string }) => ({
+    dimensionValue: value.dimensionValue ?? Value.noFilter(),
+  }),
+};
+
+const props = (inputs: Inputs<typeof meta>) => ({
+  ...inputs,
+  results: loadDataResults(inputs),
+});
+
+export const pieChartPro = {
+  Component,
+  meta,
+  preview,
+  previewConfig,
+  config: {
+    props,
+    events,
+  },
+  results: {
+    loadDataArgs: loadDataResultsArgs,
+    loadData: loadDataResults,
+  },
+} as const;
